Use a wrapper around ANW to allow updating Surface
GpuDisplaySurface are roughly created once per "display in the
guest". Prior to this change, the Surface from the viewer app would
be saved in the AndroidSurface once at initial startup. However,
when the viewer app is moved to the background, the underlying
Surface is destroyed. After this point, the Surface saved in
AndroidSurface is no longer valid. This change updates
AndroidSurface to save a pointer to a ANativeWindowWrapper struct,
which contains the latest Surface, if any, and which can be updated
by DisplayService.

Note: this does not fix the app resuming issue.

Bug: b/334903567
Test: launch ferrochrome,
      move to viewer to background,
      move viewer back to foreground
Change-Id: I647de0a7735280a757e392e878fce86f559d5a74