microdroid: Sign images using pvmfw_embedded_key
Alias microdroid_sign_key to pvmfw_embedded_key to express in code the
dependency between them, which was until now only described by a (now
superfluous) comment. As pvmfw_embedded_key is itself an alias for
avb_testkey_rsa4096, no changes are expected from this commit.

Remove the rest of the comment (about different keys) as it seems to be
more confusing than informative.

Bug: 342549834
Test: TH
Change-Id: If7ee5bacfbb2db9dd4a2369290a559c848e3a325