Ferrochrome app guides how to download the assets
Ferrochrome app which is pre-installed on the device won't have assets
(the guest OS image) in it. In that case, the app instructs the user to
follow the guideline (go/try-ferrochrome) to manually download the
version of the app with assets in it and install it.

This CL also makes some minor changes:

* Readability enhancement
   - deleted the theme with no action bar
   - increased the font size of the text view

* After the asset extraction is done, wait for 30 sec before starting
the vm launcher activity. Ideally, this shouldn't be needed but for some
unknown reason it takes some time for the filesystem to be updated after
the asset files extraction.

Bug: N/A
Test: run the app, and follow go/try-ferrochrome
Change-Id: I0ef0401067624dcfc59fefd71da9109f85cf5a6d