vmbase_example: Clarify that the payload is a BIOS
Rename all Soong targets to include "_bios" and refactor the linker
script and idmap.S to be preprocessed and receive a CPP macro
representing the "mode". Rename image.ld to reflect this.

Modify the test to use the new "_bios" binary name and refactor it to
make the core logic re-usable for future vmbase_example modes.

No functional change intended.

Test: atest vmbase_example.integration_test
Change-Id: I06679d154494d85916ff534298a066519b1b0778