virtmgr: disable disk file locking in crosvm
crosvm calls `man 2 flock` on disk image files. It used to only call it
on the "root" image, but now it calls it on all the nested images, e.g.
the "component" files in a composite disk image.

That isn't necessarily a bad thing, but the SELinux policies don't allow
it. For example, some of the disk images used by Microdroid are labelled
with the very generic "system_file" context. We should consider changing
the policies, but, in meantime, disable locking isn't a big deal.
Generally the files are either read-only at the FS level (and so the
locking is mostly redundant) or are generated per VM images (and so no
conflict is likely).

Bug: 364645148, 330911976
Test: launch microdroid
Change-Id: I5d64ee343d56620df4389c57dbbb99a57e2c3c8c